Data.Network.BGP

Description

MRT is a library for parsing Multi-Threaded Routing Toolkit (MRT) export files, of the kind you might find on the RouteViews archive.

Documentation

type ASNumber = Word32

2-byte AS number formats in BGP are unsupported

data Origin

ORIGIN is a well-known mandatory attribute that defines the origin of the path information.

See: RFC 4271, Section 5.1.1

Constructors

IGP

EGP

INCOMPLETE

data ASPath

AS_PATH is a well-known mandatory attribute that is composed of a sequence of AS path segments.

data ASPathSegment

A path segment is either an AS Set or an AS Sequence.

See RFC 4271, Section 5.1.2

Constructors

ASSet [ASNumber]

ASSequence [ASNumber]

data BGPAttributeValue

A BGP attribute consist of flags, then one of t

Constructors

Origin Origin

ASPath ASPath

NextHop ByteString

LocalPref Word32

AtomicAggregate

UnknownAttribute Word8 ByteString

data BGPAttributeFlags

data BGPAttribute

Constructors

BGPAttribute BGPAttributeFlags BGPAttributeValue