Data.Network.MRT

Description

MRT is a library for parsing Multi-Threaded Routing Toolkit (MRT) export files, of the kind you might find on the RouteViews archive.

Documentation

type Timestamp = Word32

The Timestamp type alias represents a BGP timestamp attribute, recorded as seconds since the Unix epoch.

type ASNumber = Word32

2-byte AS number formats in BGP are unsupported

data ASPathSegment

A path segment is either an AS Set or an AS Sequence.

See RFC 4271, Section 5.1.2

data BGPAttributeValue

A BGP attribute consist of flags, then one of t

Constructors

Origin Origin

ASPath ASPath

NextHop ByteString

LocalPref Word32

AtomicAggregate

UnknownAttribute Word8 ByteString

data BGPAttributeFlags

data BGPAttribute

Constructors

BGPAttribute BGPAttributeFlags BGPAttributeValue

data RIBEntry

getPeerIndex :: RIBEntry -> Word16

getOriginationTime :: RIBEntry -> Timestamp

getBGPAttributes :: RIBEntry -> [BGPAttribute]

data MRTRecord

Constructors

TableDumpV2

Fields

getSequenceNo :: Word32

getPrefix :: IPRange

getRIBEntries :: [RIBEntry]

Other Word16 Word16 Word32

data MRTMessage

getMessageTimestamp :: MRTMessage -> Timestamp

getRecord :: MRTMessage -> MRTRecord

readMessages :: ByteString -> [MRTMessage]

data IPRange :: *

A unified data for AddrRange IPv4 and AddrRange IPv6. To create this, use read "192.0.2.0/24" :: IPRange. Also, "192.0.2.0/24" can be used as literal with OverloadedStrings.

>>> (read "192.0.2.1/24" :: IPRange) == IPv4Range (read "192.0.2.0/24" :: AddrRange IPv4)
True
>>> (read "2001:db8:00:00:00:00:00:01/48" :: IPRange) == IPv6Range (read "2001:db8:00:00:00:00:00:01/48" :: AddrRange IPv6)
True